requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.34.0,<1
openai>=2.14.0,<3
azure-cosmos>=4.9.0,<5
azure-core>=1.37.0,<2
python-dotenv>=1.2.1,<2
pydantic>=2.0.0,<3
pydantic-settings>=2.0.0,<3
